requirements.txt
a2wsgi==1.10.8
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.9.1
aiohttp-retry==2.9.1
aiologic==0.14.0
aiosignal==1.3.1
aiosqlite==0.21.0
alembic==1.13.1
amqp==5.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
apache-airflow==2.8.1
apache-airflow-core==3.0.1
apache-airflow-providers-common-compat==1.7.0
apache-airflow-providers-common-io==1.2.0
apache-airflow-providers-common-sql==1.10.0
apache-airflow-providers-docker==4.4.0
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-google==10.1.0
apache-airflow-providers-http==4.8.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-smtp==2.1.0
apache-airflow-providers-sqlite==3.7.0
apache-airflow-providers-standard==1.2.0
apache-airflow-task-sdk==1.0.1
apispec==6.4.0
appdirs==1.4.4
argcomplete==3.2.1
asgiref==3.7.2
asttokens==3.0.0
asyncssh==2.21.0
atpublic==6.0.1
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
billiard==4.2.1
blinker==1.7.0
cachelib==0.9.0
cachetools==5.5.2
cadwyn==5.3.3
cattrs==24.1.3
celery==5.5.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
cloudpickle==3.1.1
colorama==0.4.6
colorlog==4.8.0
comm==0.2.2
configobj==5.0.9
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.2
cron-descriptor==1.4.0
croniter==2.0.1
cryptography==41.0.7
cycler==0.12.1
databricks-sdk==0.54.0
db-dtypes==1.4.3
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.14
dictdiffer==0.9.0
dill==0.3.1.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.4.2
docker==7.1.0
docstring_parser==0.16
docutils==0.20.1
dpath==2.2.0
dulwich==0.22.8
dvc==3.59.2
dvc-data==3.16.10
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
email-validator==1.3.1
entrypoints==0.4
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
Flask==2.2.5
Flask-AppBuilder==4.3.10
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.58.0
frozenlist==1.4.1
fsspec==2023.12.2
funcy==2.0
gcloud-aio-auth==4.2.3
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ads==26.1.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.94.0
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-automl==2.16.3
google-cloud-bigquery==3.33.0
google-cloud-bigquery-datatransfer==3.19.1
google-cloud-bigtable==2.31.0
google-cloud-build==3.31.1
google-cloud-compute==1.30.0
google-cloud-container==2.57.0
google-cloud-core==2.4.3
google-cloud-datacatalog==3.27.1
google-cloud-dataflow-client==0.9.0
google-cloud-dataform==0.6.2
google-cloud-dataplex==2.10.2
google-cloud-dataproc==5.18.1
google-cloud-dataproc-metastore==1.18.2
google-cloud-dlp==3.29.0
google-cloud-kms==3.5.1
google-cloud-language==2.17.1
google-cloud-logging==3.12.1
google-cloud-memcache==1.12.1
google-cloud-monitoring==2.27.1
google-cloud-orchestration-airflow==1.17.5
google-cloud-os-login==2.17.1
google-cloud-pubsub==2.29.0
google-cloud-redis==2.18.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-spanner==3.54.0
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-tasks==2.19.2
google-cloud-texttospeech==2.27.0
google-cloud-translate==3.20.2
google-cloud-videointelligence==2.16.1
google-cloud-vision==3.10.1
google-cloud-workflows==1.18.1
google-crc32c==1.7.1
google-genai==1.16.1
google-re2==1.1
google-resumable-media==2.7.2
googleapis-common-protos==1.62.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.0.3
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.60.0
grpcio-gcp==0.2.2
grpcio-status==1.71.0
gto==1.7.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httptools==0.6.4
httpx==0.23.3
hydra-core==1.3.2
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflection==0.5.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
iterative-telemetry==0.0.10
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.3
joblib==1.5.1
json-merge-patch==0.3.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.8.0
kiwisolver==1.4.8
kombu==5.5.3
lazy-object-proxy==1.10.0
libcst==1.7.0
limits==3.7.0
linkify-it-py==2.0.2
lockfile==0.12.2
looker-sdk==25.8.0
Mako==1.3.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.0
mdurl==0.1.2
methodtools==0.4.7
mlflow==2.22.0
mlflow-skinny==2.22.0
more-itertools==10.7.0
msgspec==0.19.0
multidict==6.0.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.2.2
omegaconf==2.3.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
ordered-set==4.1.0
orjson==3.10.18
packaging==23.2
pandas==2.2.3
pandas-gbq==0.29.0
parso==0.8.4
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.3.0
prison==0.2.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.2
psutil==5.9.7
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydata-google-auth==1.9.1
pydot==4.0.0
pygit2==1.18.0
Pygments==2.17.2
pygtrie==2.5.0
PyJWT==2.8.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.1.0
python-multipart==0.0.20
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.4.0
referencing==0.32.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retryhttp==1.3.3
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.7.0
rich-argparse==1.4.0
rich-toolkit==0.14.6
rpds-py==0.17.1
rsa==4.9.1
ruamel.yaml==0.18.11
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.15.3
scmrepo==3.3.11
semver==3.0.4
setproctitle==1.3.3
shapely==2.1.1
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.0
SQLAlchemy==1.4.51
sqlalchemy-bigquery==1.14.1
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
sqltrie==0.11.2
stack-data==0.6.3
starlette==0.46.2
structlog==25.3.0
svcs==25.1.0
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.6.0
time-machine==2.13.0
tomlkit==0.13.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
types-requests==2.32.0.20250515
typing-inspection==0.4.1
typing_extensions==4.9.0
tzdata==2023.4
uc-micro-py==1.0.2
unicodecsv==0.14.1
universal-pathlib==0.1.4
uritemplate==4.1.1
urllib3==2.0.7
uuid6==2024.7.10
uvicorn==0.34.2
uvloop==0.21.0
vine==5.1.0
voluptuous==0.15.2
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.17.0
